Implemented Merge Sort
In the sorting directory, we will implement the sorting algorithms. For now, Merge Sort has been implemented.

In the resources directory, we will input files that will be different in terms of how many numbers and how big the numbers are to test how well each sorting algorithm performs in different scenarios.

The DataReader will be used to scan the input files.